Adds more type hints to foozzer.py
Making foozzer.py work with mypy --strict is probably
not worth it since the plugin discovery with pkgutil
results in an excessive number of required type hints
that do barely add any value.
Therefore the hints in this file are considered
adequate at this point.
The plugins could do with additional type hints, though.